+++ packages/composable/src/programmaticOrders/index.ts
@@ -1,4 +1,4 @@
 export { ProgrammaticOrderApi } from './ProgrammaticOrderApi'
-export { ProgrammaticOrderApiError } from './common/types'
-export type { ProgrammaticOrderApiOptions, ProgrammaticOrderStatus } from './common/types'
-export type { GetTwapOrdersParams, TwapOrder, TwapPartOrder, TwapPartOrderStatus, TwapSchedule } from './twap/types'
+export { ProgrammaticOrderApiError } from './types'
+export type { ProgrammaticOrderApiOptions, ProgrammaticOrderStatus } from './types'
+export type { GetTwapOrdersParams, TwapOrder, TwapPartOrder, TwapPartOrderStatus, TwapSchedule } from './twap-types'
